requirements.txt
asgiref==3.2.10
astroid==2.2.5
autopep8==1.4.4
boto3==1.14.3
botocore==1.17.3
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-allauth==0.39.1
django-cors-headers==3.0.2
django-countries==5.3.3
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.9.4
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.1.0
psycopg2==2.7.5
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.3.3
six==1.12.0
sqlparse==0.3.0
stripe==2.33.0
toml==0.10.1
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==5.1.0
wrapt==1.11.2
